+++ scalafix-core/src/main/scala/scalafix/util/TokenOps.scala
@@ -1,0 +1,13 @@
+package scalafix.util
+
+import scala.meta.Token
+
+object TokenOps {
+  type TokenHash = Long
+  def hash(token: Token): TokenHash = {
+    val longHash: Long =
+      (token.productPrefix.hashCode.toLong << (62 - 8)) |
+        (token.start.toLong << (62 - (8 + 28))) | token.end
+    longHash
+  }
+}
